Share the sign-request analysis between the PSBT and transaction screens
Both approval hooks ran the same checks — payload decode, local verification,
safety analysis, MPMA recipients, structure findings, protocol context, the
not-a-Counterparty-transaction gate — written out twice, each copy carrying
comments saying it matched the other. One fix landing in one copy is all it
takes for the two screens to disagree about whether a transaction is safe.

`analyzeSignRequest` now holds that pipeline once, and both hooks call it.
`emitToBackground` was copied into all three sign hooks, differing only in a
log string; it moves to one module.

Behaviour is unchanged. The PSBT path now decodes its raw transaction before
analysing rather than after — both were already awaited serially and neither
depends on the other, and it means the output-address enrichment lands before
the safety checks read the outputs, as it did before.
