+++ src/hooks/useSignTransactionRequest.ts
@@ -12,44 +12,21 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useSearchParams } from 'react-router';
 import { parseRawTransactionLocally } from '@/core/bitcoin/localTransactionParse';
+import { fetchInputsAttachedAssets } from '@/core/counterparty/inputAssets';
 import {
-  type AttachedAssetDestination,
-  movesCounterpartyValue,
-  resolveAttachedAssetDestination,
-} from '@/core/counterparty/attachedAssetMovement';
-import {
-  fetchInputsAttachedAssets,
-  type InputAttachedAssets,
-} from '@/core/counterparty/inputAssets';
-import {
-  checkMessageStructure,
-  type StructureFinding,
-} from '@/core/counterparty/messageStructure';
-import { type ProtocolContext, resolveProtocolContext } from '@/core/counterparty/protocolContext';
-import {
-  type CounterpartyMessage,
-  decodeCounterpartyMessage,
-  describeMpmaSend,
-  fetchInputPrevouts,
-  type MpmaRecipient,
-  resolveMpmaRecipients
-} from '@/core/counterparty/transaction';
-import {
-  analyzeTransactionSafety,
-  type SafetyAnalysis,
-} from '@/core/counterparty/transactionSafety';
-import {
-  type ProviderVerificationResult, 
-  verifyProviderTransaction
-} from '@/core/counterparty/unpack';
-import type { MPMAData } from '@/core/counterparty/unpack/messages/mpma';
+  analyzeSignRequest,
+  type SignRequestAnalysis,
+} from '@/core/counterparty/signRequestAnalysis';
+import { fetchInputPrevouts } from '@/core/counterparty/transaction';
 import { extractCounterpartyPayload } from '@/core/counterparty/unpack/opReturn';
+import { emitToBackground } from '@/platform/provider/emitToBackground';
 import { getSignFlow, recordSignOutcome, type SignTransactionRequest } from '@/platform/provider/signFlow';
 
 /**
- * Decoded transaction details
+ * Decoded transaction details: what the bytes say, plus the safety analysis every signing request
+ * gets (`signRequestAnalysis.ts`, shared with the PSBT screen).
  */
-export interface DecodedTransactionInfo {
+export interface DecodedTransactionInfo extends SignRequestAnalysis {
   txid: string;
   inputs: Array<{
     txid: string;
@@ -72,47 +49,6 @@ export interface DecodedTransactionInfo {
   /** Transaction virtual size in vbytes (for fee rate calculation) */
   vsize?: number;
   hasOpReturn: boolean;
-  counterpartyMessage?: CounterpartyMessage;
-  /** Local verification result */
-  verification: ProviderVerificationResult;
-  /** Security analysis (dangerous types, suspicious outputs) */
-  safety: SafetyAnalysis;
-  /** Inputs whose UTXOs carry Counterparty assets, or whose lookup failed. */
-  attachedAssets: InputAttachedAssets[];
-  /**
-   * Message fields that reference this transaction and do not resolve against it — an attach
-   * naming an output that does not exist, a move naming a UTXO the transaction does not spend.
-   */
-  structureFindings: StructureFinding[];
-  /**
-   * Ledger facts the message does not carry, so the detail list can say what the transaction means
-   * rather than only what it contains — the order behind a cancel, the supply a destroy is measured
-   * against, the total a dividend costs.
-   */
-  protocolContext: ProtocolContext;
-  /** Where the assets attached to the signed inputs end up. Null when nothing attached moves. */
-  attachedAssetDestination: AttachedAssetDestination | null;
-  /**
-   * Recipients of an mpma_send, read from the local unpack. Empty for every other message type.
-   * These are carried in the payload rather than as outputs, so the approval screen has no other
-   * source for them.
-   */
-  mpmaRecipients: MpmaRecipient[];
-}
-
-/**
- * Send an event to the background script's EventEmitterService.
- * This is necessary because the popup and background have separate instances.
- */
-function emitToBackground(event: string, data: unknown): void {
-  chrome.runtime.sendMessage({
-    type: 'COMPOSE_EVENT',
-    event,
-    data
-  }).catch((error) => {
-    // Popup might be closing, which is fine
-    console.debug('Failed to emit sign transaction event to background:', error);
-  });
 }
 
 export function useSignTransactionRequest(signerAddress?: string) {
@@ -184,109 +120,26 @@ export function useSignTransactionRequest(signerAddress?: string) {
 
     const hasOpReturn = parsed.hasOpReturn;
 
-    let counterpartyMessage: CounterpartyMessage | undefined;
-    let counterpartyDataHex: string | undefined;
-
     // Resolve any Counterparty payload the outputs carry — plaintext or ARC4 OP_RETURN, or
     // bare-multisig data outputs, which produce no OP_RETURN at all. Classifying every encoding
     // here is what lets the sweep block apply regardless of how the message is carried.
     //
     // Read from the raw bytes, not from the API's rendering of them. Extracting from an API-
     // supplied script list keyed by an API-supplied txid would leave both sides of the comparison
-    // below rooted in the same source, so agreement would prove nothing about the bytes.
-    counterpartyDataHex = extractCounterpartyPayload(rawTxHex) ?? undefined;
+    // rooted in the same source, so agreement would prove nothing about the bytes.
+    const counterpartyDataHex = extractCounterpartyPayload(rawTxHex) ?? undefined;
 
-    // If the outputs carried Counterparty data, try API unpack for rich message info
-    if (counterpartyDataHex) {
-      try {
-        const msg = await decodeCounterpartyMessage(counterpartyDataHex);
-        if (msg) {
-          counterpartyMessage = msg;
-        }
-      } catch (err) {
-        console.warn('Failed to decode Counterparty message:', err);
-      }
-    }
-
-    // Verify locally (compares local binary unpack against API result)
-    const verification = verifyProviderTransaction(counterpartyDataHex, counterpartyMessage);
-
-    // Analyze for security risks (dangerous types, suspicious outputs)
-    const messageType = counterpartyMessage?.messageType
-      ?? verification.localUnpack?.messageType;
-    const safety = analyzeTransactionSafety(messageType, outputs, signerAddress || '');
-
-    // mpma_send is described from the bytes, not from the API's rendering of them — see
-    // resolveMpmaRecipients for why the API cannot be used for this type.
-    let mpmaRecipients: MpmaRecipient[] = [];
-    if (verification.localUnpack?.messageType === 'mpma_send' && verification.localUnpack.data) {
-      const sends = (verification.localUnpack.data as MPMAData).sends ?? [];
-      if (sends.length > 0) {
-        mpmaRecipients = await resolveMpmaRecipients(sends);
-        if (counterpartyMessage) {
-          counterpartyMessage = {
-            ...counterpartyMessage,
-            description: describeMpmaSend(mpmaRecipients),
-          };
-        }
-      }
-    }
-
-    // Independent of both decoders: does the message's own account of this transaction hold up
-    // against the transaction? Uses the local decode, since the point is to test the bytes.
-    const structureFindings = checkMessageStructure(
-      verification.localUnpack?.messageType,
-      verification.localUnpack?.data,
-      { inputs, outputs }
-    );
-
-    const { context: protocolContext, warnings: policyWarnings } = await resolveProtocolContext({
-      messageType: verification.localUnpack?.messageType,
-      data: verification.localUnpack?.data,
-      transactionId: parsed.txid,
-      apiMessageData: counterpartyMessage?.messageData,
+    // Every input is being signed here, so all of them count as sources for attached assets — a
+    // raw transaction can spend an attached UTXO just as a PSBT can, and does so with no message.
+    const analysis = await analyzeSignRequest({
+      counterpartyDataHex,
+      inputs,
       outputs,
       signerAddresses: signerAddress ? [signerAddress] : [],
-      spentUtxos: inputs.map((input) => `${input.txid}:${input.vout}`),
+      signedInputIndices: inputs.map((_, index) => index),
+      transactionId: parsed.txid,
+      attachedAssets: attachedAssetsPromise,
     });
-
-    // Policy blocks come from the same lookups as the detail list — an oracle-priced dispenser is
-    // only visible once the dispensers behind the paid address have been read.
-    if (policyWarnings.length > 0) {
-      safety.warnings = [...policyWarnings, ...safety.warnings];
-      safety.blocked = safety.blocked || policyWarnings.some((w) => w.severity === 'block');
-    }
-
-    const attachedAssets = await attachedAssetsPromise;
-
-    // The gate: a transaction this wallet signs on a site's behalf either carries a Counterparty
-    // message or spends an input carrying attached assets. Anything else is a plain Bitcoin
-    // transaction, which a site has no Counterparty reason to ask this wallet for and which the
-    // user can make in the wallet directly. Both halves are required — a message alone would miss
-    // an attached UTXO being spent alongside it, and attached assets alone miss every ordinary send.
-    if (!movesCounterpartyValue(Boolean(counterpartyDataHex), attachedAssets, inputs.map((_, index) => index))) {
-      safety.warnings = [
-        {
-          severity: 'block',
-          title: 'Blocked: Not a Counterparty Transaction',
-          message:
-            'This carries no Counterparty message and spends nothing holding Counterparty assets, ' +
-            'so signing it would move only bitcoin at a site’s direction. Make plain Bitcoin ' +
-            'payments in the wallet, where you choose the destination.',
-        },
-        ...safety.warnings,
-      ];
-      safety.blocked = true;
-    }
-
-    // A raw transaction can spend an attached UTXO just as a PSBT can, and does so with no
-    // message. Every input is being signed here, so all of them count as sources.
-    const attachedAssetDestination = resolveAttachedAssetDestination(
-      outputs,
-      attachedAssets,
-      inputs.map((_, index) => index),
-      signerAddress ? [signerAddress] : []
-    );
 
     return {
       txid: parsed.txid,
@@ -297,14 +150,7 @@ export function useSignTransactionRequest(signerAddress?: string) {
       fee,
       vsize: parsed.vsize,
       hasOpReturn,
-      counterpartyMessage,
-      verification,
-      safety,
-      attachedAssets,
-      structureFindings,
-      protocolContext,
-      attachedAssetDestination,
-      mpmaRecipients,
+      ...analysis,
     };
   }, []);
